Dependencies update and following minor changes
