+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
     ext {
-        springBootVersion = '2.1.3.RELEASE'
+        springBootVersion = '2.2.0.RELEASE'
     }
     repositories {
         mavenCentral()
@@ -17,14 +17,14 @@ buildscript {
 plugins {
     id 'application'
     id 'checkstyle'
-    id 'io.spring.dependency-management' version '1.0.6.RELEASE'
-    id 'com.github.ben-manes.versions' version '0.17.0'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+    id 'com.github.ben-manes.versions' version '0.27.0'
     id 'jacoco'
     id 'net.ltgt.apt' version '0.10'
     id 'io.franzbecker.gradle-lombok' version '1.14'
-    id 'org.owasp.dependencycheck' version '5.1.0'
+    id 'org.owasp.dependencycheck' version '5.2.2'
     id 'org.sonarqube' version '2.7.1'
-    id 'org.springframework.boot' version '2.1.3.RELEASE'
+    id 'org.springframework.boot' version '2.2.0.RELEASE'
     id 'pmd'
     id "info.solidsoft.pitest" version "1.3.0"
     id 'au.com.dius.pact' version '3.5.24'
@@ -127,7 +127,6 @@ def versions = [
         lombokVersion           : '1.16.18',
         springfoxSwagger        : '2.9.2',
         commonsIo               : '2.5',
-        feignHttpClient         : '9.5.1',
         wiremockVersion         : '2.25.0',
         reformsJavaLogging      : '5.0.1',
         serviceTokenGenerator   : '3.0.0',
@@ -136,47 +135,21 @@ def versions = [
         junit                   : '4.12',
         nimbus                  : '5.1',
         pdfbox                  : '2.0.8',
-        mockito                 : '2.8.9',
+        mockito                     : '2.8.9',
         powerMock               : '1.7.3',
         puppyCrawl              : '7.6',
         restAssured             : '3.0.3',
         serenity                : '2.0.11',
         serenityCucumber        : '1.1.3',
-        springBoot              : '2.1.3.RELEASE',
-        springOpenFeign         : '2.0.0.RELEASE',
-        springHateoas           : '0.24.0.RELEASE',
-        unirest                 : '1.4.9',
+        springBoot              : '2.2.0.RELEASE',
         pitest                  : '1.3.2',
         gradlePitest            : '1.4.0',
         sonarPitest             : '0.5',
-        ccdClient               : '4.3.1',
-        commonsBeanUtils                : '1.9.4',
-        pact_version                    : '3.5.24'
+        ccdClient               : '4.6.4',
+        commonsBeanUtils        : '1.9.4',
+        pact_version            : '3.5.24'
 
 ]
-
-dependencyManagement {
-    dependencies {
-        // CVE-2019-0232, CVE-2019-0199 - command line injections on windows
-        dependencySet(group: 'org.apache.tomcat.embed', version: '9.0.19') {
-            entry 'tomcat-embed-core'
-            entry 'tomcat-embed-el'
-            entry 'tomcat-embed-websocket'
-        }
-
-        // solves CVE-2019-12086
-        // remove once spring manager incorporates this changes
-        dependencySet(
-                group: 'com.fasterxml.jackson.core',
-                version: '2.9.9'
-                //version: dependencyManagement.importedProperties['jackson.version'] solves CVE-2019-12086
-        ) {
-            entry 'jackson-core'
-            entry 'jackson-databind'
-        }
-        dependency group: 'commons-beanutils', name: 'commons-beanutils', version: '1.9.4'
-    }
-}
 
 dependencies {
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: versions.springBoot
@@ -188,8 +161,8 @@ dependencies {
     compile group: 'commons-io', name: 'commons-io', version: versions.commonsIo
     compile(group: 'io.springfox', name: 'springfox-swagger2', version: versions.springfoxSwagger)
     compile (group: 'io.springfox', name: 'springfox-swagger-ui', version: versions.springfoxSwagger)
-    compile group: 'uk.gov.hmcts.reform', name: 'core-case-data-store-client', version: '4.6.0'
-    compile (group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign', version: '2.0.1.RELEASE')
+    compile group: 'uk.gov.hmcts.reform', name: 'core-case-data-store-client', version: versions.ccdClient
+    compile (group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign', version: '2.1.3.RELEASE')
             {
                 exclude group: 'io.reactivex', module: 'io.reactivex'
                 exclude group: 'io.reactivex', module: 'rxnetty'
@@ -198,9 +171,7 @@ dependencies {
             }
 
     compile group: 'uk.gov.hmcts.reform', name: 'service-auth-provider-client', version: versions.serviceTokenGenerator
-    compile group: 'io.github.openfeign', name: 'feign-httpclient', version: versions.feignHttpClient
     compile group: 'org.apache.commons', name: 'commons-lang3', version: versions.commonsLang3
-    compile group: 'uk.gov.hmcts.reform', name: 'core-case-data-store-client', version: versions.ccdClient
     compile group: 'uk.gov.hmcts.reform', name: 'health-spring-boot-starter', version:'0.0.5'
     compile group: 'uk.gov.hmcts.reform', name: 'properties-volume-spring-boot-starter', version:'0.0.4'
 
@@ -228,10 +199,10 @@ dependencies {
     testSmokeCompile sourceSets.test.runtimeClasspath
 
     functionalCompile group: 'junit', name: 'junit', version: versions.junit
-    functionalCompile group: 'net.serenity-bdd', name: 'serenity-core', version: '2.0.11'
-    functionalCompile group: 'net.serenity-bdd', name: 'serenity-junit', version: '2.0.11'
-    functionalCompile group: 'net.serenity-bdd', name: 'serenity-rest-assured', version: '2.0.11'
-    functionalCompile group: 'net.serenity-bdd', name: 'serenity-spring', version: '2.0.11'
+    functionalCompile group: 'net.serenity-bdd', name: 'serenity-core', version: versions.serenity
+    functionalCompile group: 'net.serenity-bdd', name: 'serenity-junit', version: versions.serenity
+    functionalCompile group: 'net.serenity-bdd', name: 'serenity-rest-assured', version: versions.serenity
+    functionalCompile group: 'net.serenity-bdd', name: 'serenity-spring', version: versions.serenity
     functionalCompile group: 'io.jsonwebtoken', name: 'jjwt', version: '0.9.0'
     functionalCompile group: 'uk.gov.hmcts.reform', name: 'service-auth-provider-client', version: '1.0.4'
     functionalCompile group: 'io.rest-assured', name: 'rest-assured', version: versions.restAssured
