+++ src/functionalTests/resources/application.properties
@@ -17,6 +17,7 @@ idam.api.url=${IDAM_URL:http://localhost:4501}
 idam.api.url.local=http://localhost:4501/testing-support/lease
 idam.api.secret=${OAUTH2_CLIENT_FINREM:AAAAAAAAAAAAAAAA}
 
+allowed-services=
 
 idam.userpassword=${PASSWORD-SOLICITOR:London01}
 idam.whitelist.url=https://localhost:9000/oauth2/callback
