+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/error/GlobalExceptionHandler.java
@@ -7,12 +7,20 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.client.HttpServerErrorException;
+import uk.gov.hmcts.reform.authorisation.exceptions.InvalidTokenException;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ApiError;
+
+import static org.springframework.http.HttpStatus.FORBIDDEN;
+import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
+import static org.springframework.http.ResponseEntity.status;
 
 @ControllerAdvice
 @Slf4j
 public class GlobalExceptionHandler {
 
     public static final String SERVER_ERROR_MSG = "Some server side exception occurred. Please check logs for details";
+    private static final ApiError API_ERROR_FORBIDDEN = new ApiError("S2S token is not authorized to use the service");
 
     @ExceptionHandler(FeignException.class)
     ResponseEntity<Object> handleFeignException(FeignException exception) {
@@ -36,5 +44,23 @@ public class GlobalExceptionHandler {
     ResponseEntity<Object> handleNoSuchFieldExistsException(NoSuchFieldExistsException exception) {
         log.error(exception.getMessage(), exception);
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(SERVER_ERROR_MSG);
+    }
+
+    @ExceptionHandler(InvalidTokenException.class)
+    protected ResponseEntity<ApiError> handleInvalidTokenException(InvalidTokenException exc) {
+        log.warn(exc.getMessage(), exc);
+        return status(UNAUTHORIZED).body(new ApiError(exc.getMessage()));
+    }
+
+    @ExceptionHandler(UnauthenticatedException.class)
+    protected ResponseEntity<ApiError> handleUnauthenticatedException(UnauthenticatedException exc) {
+        log.warn(exc.getMessage(), exc);
+        return status(UNAUTHORIZED).body(new ApiError(exc.getMessage()));
+    }
+
+    @ExceptionHandler(ForbiddenException.class)
+    protected ResponseEntity<ApiError> handleForbiddenException(ForbiddenException exc) {
+        log.warn(exc.getMessage(), exc);
+        return status(FORBIDDEN).body(API_ERROR_FORBIDDEN);
     }
 }
