+++ src/main/resources/application.properties
@@ -107,4 +107,3 @@ optionsValueFile=/options/options-id-value-transform.json
 idam.url=${IDAM_API_URL:http://localhost:4501}
 idam.api=/details
 idam.health.url=${idam.url}/health
-idam.s2s-auth.url=${idam.url}
