+++ src/test/resources/application.properties
@@ -61,3 +61,6 @@ optionsValueFile=/options/options-id-value-transform.json
 idam.url=${IDAM_API_URL:http://localhost:4501}
 idam.api=/details
 idam.health.url=${idam.url}/health
+idam.s2s-auth.url=false
+
+allowed-services=
